move fork detection to `forkdetection.go`
